Refactor pollution monitoring system in ThermalPowerPlant

+++ src/Common/Measurement.java
@@ -4,7 +4,7 @@ public class Measurement {
     public long timestamp;
     public double co2Value;
 
-    public Measurement(long timestamp, int co2Value) {
+    public Measurement(long timestamp, double co2Value) {
         this.timestamp = timestamp;
         this.co2Value = co2Value;
     }
